pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.32.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.0",
    "google-generativeai>=0.8.0"
]